apps/workers/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.2
pydantic-settings==2.1.0

# Database & Storage
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.23
alembic==1.13.1
redis[hiredis]==5.0.1
clickhouse-driver==0.2.6
boto3==1.34.0

# Message Queue & Events
nats-py==2.6.0
celery==5.3.4

# HTTP & API
httpx==0.25.2
aiohttp==3.9.1

# Security & Auth
cryptography==41.0.8
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# AI & Content Generation
openai==1.6.1
anthropic==0.7.8
langchain==0.1.0
langchain-openai==0.0.2

# Email & Communications
sendgrid==6.10.0
twilio==8.11.0
python-multipart==0.0.6

# Utilities
python-dotenv==1.0.0
structlog==23.2.0
tenacity==8.2.3
jinja2==3.1.2
markupsafe==2.1.3
pytz==2023.3
python-dateutil==2.8.2

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Monitoring & Observability
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
sentry-sdk[fastapi]==1.39.2
